+++ Galapagos/CrossoverOperators/Binary/NoOpBinaryCrossover.cs
@@ -18,7 +18,7 @@ namespace Galapagos.CrossoverOperators.Binary
         /// Constructs a new instance of the <see cref="NoOpBinaryCrossover"/> class.
         /// </summary>
         /// <param name="weigth">The crossover weight.</param>
-        internal NoOpBinaryCrossover(uint weigth = 1)
+        internal NoOpBinaryCrossover(double weigth = 1)
             : base(weigth) { }
     }
 }
