+++ Galapagos/MutationOperators/Binary/FlipBitMutation.cs
@@ -17,7 +17,7 @@ namespace Galapagos.MutationOperators.Binary
         /// Constructs a new instance of the <see cref="FlipBitMutation"/> class.
         /// </summary>
         /// <param name="weigth">The crossover weight.</param>
-        internal FlipBitMutation(uint weigth = 1)
+        internal FlipBitMutation(double weigth = 1)
             : base(weigth) { }
 
         /// <summary>
