+++ Galapagos/MutationOperators/Permutation/CyclicShiftPermutationMutation.cs
@@ -18,7 +18,7 @@ namespace Galapagos.MutationOperators.Permutation
         /// Constructs a new instance of the <see cref="CyclicShiftPermutationMutation"/> class.
         /// </summary>
         /// <param name="weigth">The crossover weight.</param>
-        internal CyclicShiftPermutationMutation(uint weigth = 1)
+        internal CyclicShiftPermutationMutation(double weigth = 1)
             : base(weigth) { }
 
         /// <summary>
