+++ Galapagos/MutationOperators/Shared/CyclicShiftMutation.cs
@@ -19,7 +19,7 @@ namespace Galapagos.MutationOperators.Shared
         /// Constructs a new instance of the <see cref="CyclicShiftMutation"/> class.
         /// </summary>
         /// <param name="weigth">The crossover weight.</param>
-        protected CyclicShiftMutation(uint weigth = 1)
+        protected CyclicShiftMutation(double weigth = 1)
             : base(weigth) { }
 
         /// <summary>
